+++ CHANGELOG.md
@@ -33,6 +33,7 @@ This changelog is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.
 ### Fixed
 
 - Added explicit read permissions to examples.yml (#623)
+- Added node_account_ids selection method support (#362)
 - Improved type hinting in `file_append_transaction.py` to resolve 'mypy --strict` errors. ([#495](https://github.com/hiero-ledger/hiero-sdk-python/issues/495))
 - fix: Resolve `__eq__` type conflict in `CustomFee` class (#627)